实现回复功能

+++ src/main/java/com/xiaobian/community/controller/QuestionController.java
@@ -16,7 +16,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(value = "id") Integer id,
+    public String question(@PathVariable(value = "id") Long id,
                            Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         questionService.incView(id);
